+++ CHANGELOG.md
@@ -3,6 +3,7 @@ QML Material CHANGELOG
 
  * Add a ProgressCircle component following the Google spec
  * Add a tooltip component and use it in the ActionBar
+ * [Tooltip] Clean up the code and improve the style
 
 ### Version 0.0.5 (March 4, 2015)